requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
anthropic==0.40.0
python-dotenv==1.0.1
python-multipart==0.0.12
